+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- GroupsDetailPanel: implement posts paging [#572](https://github.com/rokwire/illinois-app/issues/572).
 
 ## [2.6.8] - 2021-07-22
 ### Changed