+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+- GroupsDetailPanel: load older posts with q button [#591](https://github.com/rokwire/illinois-app/issues/591).
 
 ## [2.6.9] - 2021-07-23
 ### Changed