+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated group posts display time [#593](https://github.com/rokwire/illinois-app/issues/593).
 ### Fixed
 - Fixed replies count update [#595](https://github.com/rokwire/illinois-app/issues/595).
+- Display the right group update time [#597](https://github.com/rokwire/illinois-app/issues/597).
 
 ## [2.6.9] - 2021-07-23
 ### Changed